feat: compile CSV file in server action
